pyproject.toml
[tool.poetry]
name = "wallybaba"
version = "1.0.0"
description = "WallyBaba Token - The Wise Dog of DeFi"
authors = ["Wallatoshi Babymoto"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
solders = "^0.9.2"  # Pure Python Solana library
python-dotenv = "^1.0.0"
requests = "^2.31.0"
python-dateutil = "^2.8.2"
cryptography = "^41.0.4"
Pillow = "^10.0.0"
aiohttp = "^3.8.5"
retry = "^0.9.2"
cachetools = "^4.2.4"  # Alternative to lru-dict
websockets = "^10.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
black = "^23.9.1"
isort = "^5.12.0"
mypy = "^1.5.1"
pylint = "^3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
process-image = "src.image_processor:main"
deploy-branding = "src.deploy_branding:main"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pylint.messages_control]
disable = [
    "C0111",  # missing-docstring
    "C0103",  # invalid-name
    "R0903",  # too-few-public-methods
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-ra -q" 